+++ packages/react-magma-dom/src/components/Alert/index.tsx
@@ -34,6 +34,7 @@ export interface AlertProps extends React.HTMLAttributes<HTMLDivElement> {
   isDismissed?: boolean;
   isDismissible?: boolean;
   isInverse?: boolean;
+  isToast?: boolean;
   onDismiss?: () => void;
   ref?: any;
   testId?: string;
@@ -45,7 +46,7 @@ export const transitionDuration = 500;
 function buildAlertBackground(props) {
   switch (props.variant) {
     case 'info':
-      return props.theme.colors.neutral02;
+      return props.theme.colors.neutral01;
     case 'success':
       return props.theme.colors.success01;
     case 'warning':
@@ -53,12 +54,16 @@ function buildAlertBackground(props) {
     case 'danger':
       return props.theme.colors.danger;
     default:
-      return props.theme.colors.neutral02;
+      return props.theme.colors.neutral01;
   }
 }
 
 const StyledAlert = styled.div<AlertProps>`
   align-items: stretch;
+  animation: ${props =>
+    props.isExiting
+      ? `fadeout ${transitionDuration}ms`
+      : `fadein ${transitionDuration}ms`};
   background-color: ${props => buildAlertBackground(props)};
   border-radius: 3px;
   color: ${props =>
@@ -70,18 +75,33 @@ const StyledAlert = styled.div<AlertProps>`
   padding: 0;
   margin-bottom: 20px;
   max-width: 100%;
-  animation: ${props =>
-    props.isExiting
-      ? `fadeout ${transitionDuration}ms`
-      : `fadein ${transitionDuration}ms`};
 
   &:focus {
     outline: 2px dotted ${props =>
       props.isInverse
-        ? props.theme.colors.neutral08
+        ? props.theme.colors.focusInverse
         : props.theme.colors.focus};
     }
   }
+
+  ${props =>
+    props.isToast &&
+    css`
+      animation: ${props.isExiting
+        ? `slideout ${transitionDuration}ms`
+        : `slidein ${transitionDuration}ms`};
+      border: 1px solid ${props.theme.colors.neutral08};
+      border-radius: 5px;
+      box-shadow: 0 2px 8px 0 rgba(0, 0, 0, 0.4);
+      min-width: 375px;
+      margin: 0 auto;
+
+      @media (max-width: 600px) {
+        font-size: 13px;
+        padding-left: 15px;
+        width: 100%;
+      }
+    `}
 
   @keyframes fadein {
     from {
@@ -101,6 +121,24 @@ const StyledAlert = styled.div<AlertProps>`
     }
   }
 
+  @keyframes slidein {
+    from {
+      bottom: -500px;
+    }
+    to {
+      bottom: 0;
+    }
+  }
+
+  @keyframes slideout {
+    from {
+      bottom: -500px;
+    }
+    to {
+      bottom: -500px;
+    }
+  }
+
   a {
     color: inherit;
     font-weight: 600;
@@ -109,8 +147,8 @@ const StyledAlert = styled.div<AlertProps>`
     &:focus {
       outline: 2px dotted ${props =>
         props.variant === 'warning'
-          ? props.theme.colors.neutral01
-          : props.theme.colors.neutral08};
+          ? props.theme.colors.focus
+          : props.theme.colors.focusInverse};
       }
     }
   }
@@ -118,7 +156,7 @@ const StyledAlert = styled.div<AlertProps>`
 
 const AlertContents = styled.div`
   flex-grow: 1;
-  padding: 10px 15px 10px 0;
+  padding: 13px 15px 13px 0;
 `;
 
 const IconWrapperStyles = css`
@@ -127,9 +165,17 @@ const IconWrapperStyles = css`
   flex-shrink: 0;
 `;
 
-const IconWrapper = styled.span`
+const IconWrapper = styled.span<{ isToast?: boolean }>`
   ${IconWrapperStyles}
   padding: 0 10px 0 15px;
+
+  ${props =>
+    props.isToast &&
+    css`
+      @media (max-width: 600px) {
+        display: none;
+      }
+    `}
 `;
 
 const DismissibleIconWrapper = styled.span<AlertProps>`
@@ -161,8 +207,8 @@ const DismissButton = styled(IconButton, { shouldForwardProp })<{
     outline: 2px dotted
       ${({ alertVariant, theme }) =>
         alertVariant === 'warning'
-          ? theme.colors.neutral01
-          : theme.colors.neutral08};
+          ? theme.colors.focus
+          : theme.colors.focusInverse};
     outline-offset: 0 !important;
   }
 
@@ -171,8 +217,8 @@ const DismissButton = styled(IconButton, { shouldForwardProp })<{
     :not(:disabled):before {
       background: ${({ alertVariant, theme }) =>
         alertVariant === 'warning'
-          ? theme.colors.neutral01
-          : theme.colors.neutral08};
+          ? theme.colors.focus
+          : theme.colors.focusInverse};
       opacity: 0.15;
     }
 
@@ -182,11 +228,11 @@ const DismissButton = styled(IconButton, { shouldForwardProp })<{
   }
 `;
 
-function renderIcon(variant = 'info') {
+function renderIcon(variant = 'info', isToast?: boolean) {
   const Icon = VARIANT_ICON[variant];
 
   return (
-    <IconWrapper>
+    <IconWrapper isToast>
       <Icon size={20} />
     </IconWrapper>
   );
@@ -205,6 +251,7 @@ export const Alert: React.FunctionComponent<AlertProps> = React.forwardRef(
       isDismissible,
       isExiting: externalIsExiting,
       isInverse,
+      isToast,
       onDismiss,
       ...other
     }: AlertProps,
@@ -243,10 +290,11 @@ export const Alert: React.FunctionComponent<AlertProps> = React.forwardRef(
         tabIndex={-1}
         isInverse={isInverse}
         isExiting={isExiting}
+        isToast={isToast}
         variant={variant}
         theme={theme}
       >
-        {renderIcon(variant)}
+        {renderIcon(variant, isToast)}
         <AlertContents>{children}</AlertContents>
         {isDismissible && (
           <DismissibleIconWrapper variant={variant} theme={theme}>