+++ CHANGELOG.md
@@ -3,6 +3,59 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.34](http://stash.corp.web:7999/FRONT/react-magma/compare/v2.0.0-beta.33...v2.0.0-beta.34) (2020-04-09)
+
+
+### Bug Fixes
+
+* **search:** add and document style and testid props ([ce089c6](http://stash.corp.web:7999/FRONT/react-magma/commits/ce089c6))
+* **tabs:** fix TabsProps interface ([acd9e41](http://stash.corp.web:7999/FRONT/react-magma/commits/acd9e41))
+
+
+
+
+
+# [2.0.0-beta.33](http://stash.corp.web:7999/FRONT/react-magma/compare/v2.0.0-beta.32...v2.0.0-beta.33) (2020-04-08)
+
+
+### Features
+
+* **toast:** upadted animation, mobile styles ([96e4519](http://stash.corp.web:7999/FRONT/react-magma/commits/96e4519))
+* **toasts:** update styles based on orion feedback ([79f6660](http://stash.corp.web:7999/FRONT/react-magma/commits/79f6660))
+
+
+
+
+
+# [2.0.0-beta.32](http://stash.corp.web:7999/FRONT/react-magma/compare/v2.0.0-beta.31...v2.0.0-beta.32) (2020-04-07)
+
+
+### Features
+
+* **calendar,search icon:** update calendar icon, update icon sizes ([444c08e](http://stash.corp.web:7999/FRONT/react-magma/commits/444c08e))
+* **focus:** update color ([7bd022c](http://stash.corp.web:7999/FRONT/react-magma/commits/7bd022c))
+* **icons:** new icons -- open envelopes and signout ([4d8c2e7](http://stash.corp.web:7999/FRONT/react-magma/commits/4d8c2e7))
+
+
+
+
+
+# [2.0.0-beta.31](http://stash.corp.web:7999/FRONT/react-magma/compare/v2.0.0-beta.30...v2.0.0-beta.31) (2020-04-03)
+
+**Note:** Version bump only for package react-magma
+
+
+
+
+
+# [2.0.0-beta.30](http://stash.corp.web:7999/FRONT/react-magma/compare/v2.0.0-beta.29...v2.0.0-beta.30) (2020-04-01)
+
+**Note:** Version bump only for package react-magma
+
+
+
+
+
 # [2.0.0-beta.29](http://stash.corp.web:7999/FRONT/react-magma/compare/v2.0.0-beta.28...v2.0.0-beta.29) (2020-03-30)